API TaskUser,TaskDepartment,TaskHistory,Department,DepartmentUser
insert API TaskUser,TaskDepartment,TaskHistory,Department,DepartmentUser;Fix API TaskUser

+++ server/Interfaces/IRepository.cs
@@ -6,7 +6,8 @@ public interface IRepository<T> where T : class
 {
     IEnumerable<T> GetAll(Expression<Func<T, bool>>? filter = null, string includeProperties = null);
     T Get(Expression<Func<T, bool>>? filter, string includeProperties = null);
-    T GetById(int id);
+    T GetById(long id);
+    
     T Add(T entity);
     bool Any(Expression<Func<T, bool>> filter);
     T Remove(T entity);
